observers extracted to a separate manager to be able to test
