+++ app/src/test/kotlin/com/wire/android/ui/WireActivityViewModelTest.kt
@@ -27,22 +27,16 @@ import com.wire.android.config.mockUri
 import com.wire.android.di.AuthServerConfigProvider
 import com.wire.android.di.ObserveSyncStateUseCaseProvider
 import com.wire.android.feature.AccountSwitchUseCase
-import com.wire.android.framework.TestUser
 import com.wire.android.migration.MigrationManager
 import com.wire.android.navigation.BackStackMode
 import com.wire.android.navigation.NavigationCommand
 import com.wire.android.navigation.NavigationItem
 import com.wire.android.navigation.NavigationManager
-import com.wire.android.notification.NotificationChannelsManager
-import com.wire.android.notification.WireNotificationManager
-import com.wire.android.services.ServicesManager
 import com.wire.android.util.deeplink.DeepLinkProcessor
 import com.wire.android.util.deeplink.DeepLinkResult
 import com.wire.android.util.newServerConfig
 import com.wire.kalium.logic.data.id.ConversationId
 import com.wire.kalium.logic.data.id.QualifiedID
-import com.wire.kalium.logic.data.team.Team
-import com.wire.kalium.logic.data.user.SelfUser
 import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.feature.appVersioning.ObserveIfAppUpdateRequiredUseCase
 import com.wire.kalium.logic.feature.auth.AccountInfo
@@ -51,8 +45,6 @@ import com.wire.kalium.logic.feature.server.GetServerConfigUseCase
 import com.wire.kalium.logic.feature.session.CurrentSessionFlowUseCase
 import com.wire.kalium.logic.feature.session.CurrentSessionResult
 import com.wire.kalium.logic.feature.session.GetSessionsUseCase
-import com.wire.kalium.logic.feature.user.ObserveValidAccountsUseCase
-import com.wire.kalium.logic.feature.user.webSocketStatus.ObservePersistentWebSocketConnectionStatusUseCase
 import com.wire.kalium.logic.sync.ObserveSyncStateUseCase
 import io.mockk.MockKAnnotations
 import io.mockk.coEvery
@@ -388,21 +380,6 @@ class WireActivityViewModelTest {
         assertEquals(true, viewModel.globalAppState.updateAppDialog)
     }
 
-    @Test
-    fun `given few valid accounts, then notificationChannels creating is called`() {
-        val accs = listOf(
-            TestUser.SELF_USER,
-            TestUser.SELF_USER.copy(id = TestUser.USER_ID.copy(value = "something else"))
-        )
-        val (arrangement, _) = Arrangement()
-            .withSomeCurrentSession()
-            .withValidAccounts(accs.map { it to null })
-            .arrange()
-
-        coVerify(exactly = 1) { arrangement.notificationChannelsManager.createNotificationChannels(listOf()) }
-        coVerify(exactly = 1) { arrangement.notificationChannelsManager.createNotificationChannels(accs) }
-    }
-
     private class Arrangement {
         init {
             // Tests setup
@@ -413,19 +390,12 @@ class WireActivityViewModelTest {
             coEvery { currentSessionFlow() } returns flowOf()
             coEvery { getServerConfigUseCase(any()) } returns GetServerConfigResult.Success(newServerConfig(1).links)
             coEvery { deepLinkProcessor(any(), any()) } returns DeepLinkResult.Unknown
-            coEvery { notificationManager.observeNotificationsAndCallsWhileRunning(any(), any(), any()) } returns Unit
             coEvery { navigationManager.navigate(any()) } returns Unit
-            coEvery { observePersistentWebSocketConnectionStatus() } returns
-                    ObservePersistentWebSocketConnectionStatusUseCase.Result.Success(
-                        flowOf(listOf())
-                    )
             coEvery { getSessionsUseCase.invoke() }
             coEvery { migrationManager.shouldMigrate() } returns false
             every { observeSyncStateUseCaseProviderFactory.create(any()).observeSyncState } returns observeSyncStateUseCase
             every { observeSyncStateUseCase() } returns emptyFlow()
             coEvery { observeIfAppUpdateRequired(any()) } returns flowOf(false)
-            every { notificationChannelsManager.createNotificationChannels(any()) } returns Unit
-            coEvery { observeValidAccounts() } returns flowOf(listOf())
         }
 
         @MockK
@@ -438,13 +408,7 @@ class WireActivityViewModelTest {
         lateinit var deepLinkProcessor: DeepLinkProcessor
 
         @MockK
-        lateinit var notificationManager: WireNotificationManager
-
-        @MockK
         lateinit var navigationManager: NavigationManager
-
-        @MockK
-        lateinit var observePersistentWebSocketConnectionStatus: ObservePersistentWebSocketConnectionStatusUseCase
 
         @MockK
         lateinit var getSessionsUseCase: GetSessionsUseCase
@@ -465,16 +429,7 @@ class WireActivityViewModelTest {
         private lateinit var observeSyncStateUseCaseProviderFactory: ObserveSyncStateUseCaseProvider.Factory
 
         @MockK
-        lateinit var servicesManager: ServicesManager
-
-        @MockK
         lateinit var observeIfAppUpdateRequired: ObserveIfAppUpdateRequiredUseCase
-
-        @MockK
-        lateinit var observeValidAccounts: ObserveValidAccountsUseCase
-
-        @MockK
-        lateinit var notificationChannelsManager: NotificationChannelsManager
 
         private val viewModel by lazy {
             WireActivityViewModel(
@@ -482,18 +437,13 @@ class WireActivityViewModelTest {
                 currentSessionFlow = currentSessionFlow,
                 getServerConfigUseCase = getServerConfigUseCase,
                 deepLinkProcessor = deepLinkProcessor,
-                notificationManager = notificationManager,
                 navigationManager = navigationManager,
                 authServerConfigProvider = authServerConfigProvider,
-                observePersistentWebSocketConnectionStatus = observePersistentWebSocketConnectionStatus,
                 getSessions = getSessionsUseCase,
                 accountSwitch = switchAccount,
                 migrationManager = migrationManager,
                 observeSyncStateUseCaseProviderFactory = observeSyncStateUseCaseProviderFactory,
-                servicesManager = servicesManager,
-                observeIfAppUpdateRequired = observeIfAppUpdateRequired,
-                observeValidAccounts = observeValidAccounts,
-                notificationChannelsManager = notificationChannelsManager
+                observeIfAppUpdateRequired = observeIfAppUpdateRequired
             )
         }
 
@@ -514,10 +464,6 @@ class WireActivityViewModelTest {
 
         fun withAppUpdateRequired(result: Boolean): Arrangement = apply {
             coEvery { observeIfAppUpdateRequired(any()) } returns flowOf(result)
-        }
-
-        fun withValidAccounts(list: List<Pair<SelfUser, Team?>>): Arrangement = apply {
-            coEvery { observeValidAccounts() } returns flowOf(list)
         }
 
         fun arrange() = this to viewModel
